requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Deprecated==1.2.14
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
microsoft-kiota-abstractions==1.4.6
microsoft-kiota-authentication-azure==1.4.6
microsoft-kiota-http==1.4.6
microsoft-kiota-serialization-form==1.4.6
microsoft-kiota-serialization-json==1.4.6
microsoft-kiota-serialization-multipart==1.4.6
microsoft-kiota-serialization-text==1.4.6
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.1.6
msgraph-sdk==1.11.0
multidict==6.1.0
Office365-REST-Python-Client==2.5.13
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
pendulum==3.0.0
portalocker==2.10.1
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
std-uritemplate==1.0.6
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.15.5
zipp==3.20.2
